prevent overwriting existing window.onload functions
If another JavaScript already defines a window.onload function,
including the `github_issue_links.js` script will overwrite that
function.
To prevent this, save the function that is already defined (if there
is none, it'll just be empty, so no problem) and add it back as
part of the function we're defining.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>