Revert fd4714173449744ee3db85272cdbee94079ceed5
The gh-jira-sync-bot app is now enabled on all repositories under
the https://github.com/canonical organisation. This is a recent change
was not the case when we this app was enabled for the starter pack.

We do not want all the forked repositories to automatically sync
to the documentation Jira board.