init script: create link for .wokeignore
When running woke locally, the .wokeignore file must be in the
folder where you run it (for example, the `docs` folder).
When running on GitHub, we run woke from the root folder though,
and it cannot deal with a .wokeignore file in a subdirectory.
Therefore, have the script create a symlink (unless there already
is a .wokeignore file in the root folder).

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>