Restructure admin and user information
Make the readme be a landing page that describes the project and offers
two paths for the user. Two new pages were added:

* Enable the starter pack (help-enable)
* Use the stater pack (help-usage)

Keep the existing content as intact as possible by moving sections to
the appropriate page. Content modifications include:

1. The top of both new pages needed some work.
2. A small section on building was added to help-enable.
3. The headings' underlining needed to be changed throughout
   help-enable.

A subsequent PR will improve the wording of both new pages and the index
page.