Orphan `setup-script`
Deal with poor GitHub support for RST.

The setup-script is currently a sub-page to the README, according to a
conscious decision. This required a `toctree` directive in the README.
However, in GitHub the latter gets exposed in the rendered HTML.

For now, orphan the setup-script page until we decide what to do. As-is,
the README's appearance in GitHub is intolerable. The fact that this
script is considered beta at this time lessens the impact of this change.

Also fix the link to setup-script in the body of the README.