document sequential_nav in the readme
Document how you can override global configuration through page-wide
metadata. The main use case for this is sequential_nav (implemented
in our theme), but it's also relevant for hide-toc (from Furo).

I'm not aware of other settings that can be overridden, but we can
add them to the section if we find others.

Closes #170

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>