change build job for build_requirements.py
Changing the build job to `pre_install` for the `build_requirements.py`
script means that the build machine will already be fully set up, so
that you can use pip to install any packages that are required in
`custom_conf.py` and will otherwise give errors.

This is required to work around #197.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>