fix: manpage url linkcheck failures
The Ubuntu manpage urls return a "404 Not Found" error, because when verified
by linkcheck the JavaScript code that will look up the searched manpage at
runtime does not get executed.

To solve this error we staticaly define the URL that would otherwise be looked
up at runtime. Therefore we use dist-info to look up the latest Ubuntu
distribution.

One drawback is that the documentation has to periodically be regenerated,
because a newer Ubuntu distribution will be released every 6 months.

Bug: #224