ci: Add check for building Sphinx venv from source
The purpose of this workflow file is to confirm that the Sphinx
virtual environment can be built from source, consequently
documenting the packages required in the build environment to do
that.

This is needed because some projects embeds the documentation into
built artifacts which involves rendering the documentation on the
target architecture.

Depending on the architecture, pip may or may not have already
built wheels available, and as such we need to make sure building
wheels from source can succeed.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>