update the header file
Update the header file to include all related resources that are
specified in the conf.py file.

The header file really lives in the canonical-sphinx extension,
but it is included here as well to make it easy to customise.
I'll open a PR to add the changes to canonical-sphinx as well.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>