Makefile: Do not use \n in echo
Using \n' in the 'echo' command is not portable.

In bash:

  $ echo 'hello\n'
  hello\n
  $

In dash:

  $ echo 'hello\n'
  hello

  $

The current code assumes the behavior of dash's built-in echo command.
This is likely the case because GNU Make uses /bin/sh as the default
shell, and /bin/sh is a symlink to dash by default on Ubuntu.

If a different shell (e.g., bash) is used, '\n' may not be interpreted
as a newline, resulting in an unreadable message.

A more portable approach is either use the 'printf' command, or split
the message across multiple 'echo' commands.

This commit adopts the latter.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>