Suggest (#376)
* Remove impossible warning

It's impossible to skip heading levels in reST (it's designed to avoid
this as several of its output formats cannot skip levels either, due to
recursive section structures).

* Correct underline definition

reST permits heading underlines to be longer than the corresponding
title, which some authors find convenient (i.e. it's fine to just use a
full line of "=")

https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#sections

* Only use admonitions advised by the style guide

The general style guide at https://docs.ubuntu.com/styleguide/en advises
under the admonitions section that only the "warning" and "information"
admonition callouts are permitted. Given "information" doesn't exist in
reST, "note" is probably the next best thing. Changed "caution" to
"warning" and removed "tip" and "important".

* Add a section on linking to the Ubuntu man-pages

* Add note about image output formats

Using * as the image extension allows sphinx to pick the "best" image
format for the output it's producing. For example, PNG for web, but
embedded PS for PDF.

* Fix md inclusion example

* Apply rkratky's suggestions from code review

Co-authored-by: Robert Krátký <robert.kratky@canonical.com>

---------

Co-authored-by: Robert Krátký <robert.kratky@canonical.com>