feat: Python3 venv detection (#384)
* docs/Makefile: do implicit venv support detection

Opportunistically try to create the venv and if it fails, hint the user to
install `python3-venv`. This avoids spawning another python3 process just for
the detection.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>

* docs/Makefile: avoid `'` in comments

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>

* docs/Makefile: prefer `command -v` to `type`

Both are shell builtins but `type` is more for human and `command -v` for shell:

```
$ command -v python3; echo $?
/usr/bin/python3
0
$ command -v python4; echo $?
1
```

```
$ type python3; echo $?
python3 is /usr/bin/python3
0
$ type python4; echo $?
bash: type: python4: not found
1
```

Also, `command -v` does not emit anything on `stderr`, it just reflect the
availability with its exit code.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>

---------

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>